feat: #1325 Update webhook page tests
