+++ packages/marketplace/src/components/pages/developer-webhooks/index.ts
@@ -1,0 +1,2 @@
+import DeveloperWebhooks from './developer-webhooks'
+export default DeveloperWebhooks
